block a non-url warning during DB cleanup for external playlists.
avoid an undef warning for $_unknownGenreID



git-svn-id: file:///home/awy/mirror/slim/trunk/server@1912 62299810-d8cb-41fd-93b7-d32162e5a4a4

+++ Slim/DataStores/DBI/DBIStore.pm
@@ -564,7 +564,12 @@ sub clearStaleEntries {
 	my $tracks = Slim::DataStores::DBI::Track->retrieve_all();
 
 	while (my $track = $tracks->next()) {
-
+	
+		unless (Slim::Music::Info::isFileURL($track->url())) {
+			undef $track;
+			next;
+		}
+		
 		my $filepath = Slim::Utils::Misc::pathFromFileURL($track->url());
 
 		# Don't use _hasChanged - because that does more than we want.
@@ -1086,7 +1091,7 @@ sub _postCheckAttributes {
 
 		my @genres = Slim::DataStores::DBI::GenreTrack->add($genre, $track);
 
-		if (!$create) {
+		if (!$create && defined $_unknownGenreID) {
 
 			foreach my $gen (@genres) {
 				$gen->delete() if ($gen->id() eq $_unknownGenreID);
